restructure the content library folders
