+++ mobile/Features/ContentLibrary/Components/ContentFakeData.jsx
@@ -1,0 +1,28 @@
+const fakeData = [
+  {
+    id: 'bd7acbea-c1b1-46c2-aed5-3ad53abb28ba',
+    title: 'resource name',
+  },
+  {
+    id: '3ac68afc-c605-48d3-a4f8-fbd91aa97f63',
+    title: 'resource name',
+  },
+  {
+    id: '58694a0f-3da1-471f-bd96-145571e29d72',
+    title: 'resource name',
+  },
+  {
+    id: 'bd7acbea-c1b1-42c2-aed5-3ad53abb28ba',
+    title: 'resource name',
+  },
+  {
+    id: '3ac68afc-c605-46d3-a4f8-fbd91aa97f63',
+    title: 'resource name',
+  },
+  {
+    id: '58694a0f-3da2-471f-bd96-145571e29d72',
+    title: 'resource name',
+  },
+];
+
+export default fakeData;
