--- mobile/Features/Other/Screens/Favorites.jsx
@@ -1,15 +1,0 @@
-// import { useState, useEffect } from 'react';
-import {
-  Text, View,
-} from 'react-native';
-
-export default function Favorites() {
-  return (
-    <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
-      <Text>Hi</Text>
-    </View>
-  );
-}
-
-Favorites.propTypes = {
-};
